# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BasketWise API",
    "description": "\nCompare an Australian grocery basket across **Coles, Woolworths, ALDI and Harris Farm**.\n\n### The three routes you need\n1. `GET /categories` \u2014 build the browse nav. Ids for requests, names for display.\n2. `GET /products` \u2014 list, filter, paginate, search.\n3. `POST /compare` \u2014 send the basket, get three buying strategies.\n\n### Rules that keep us consistent\n* The basket stores the canonical `product_id` **only** \u2014 never a retailer's SKU.\n* The frontend never totals a basket itself; `POST /compare` owns that arithmetic.\n* Category labels come from `GET /categories`, never from a local map.\n\n### Generating a TypeScript client\n```bash\nnpx openapi-typescript https://basket.taskglass.work/openapi.json -o src/api/schema.d.ts\n```\n\nNo authentication. Data is rebuilt daily; `GET /health` reports the current `build_id`.\n",
    "contact": {
      "name": "BasketWise",
      "url": "https://basket.taskglass.work/docs"
    },
    "license": {
      "name": "Internal use"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "List, filter and search canonical products",
        "description": "Bounded listing of canonical products. Returns `[]` when nothing matches -- never 404.\n\n**Load the landing page in two steps:** request a small first page (`?essential=true&limit=20`) for first paint, then page with `limit`/`offset` once the app has mounted. `limit` is hard-capped at 100 so the full 50k-product catalogue can never be pulled in one request.\n\nFilters combine with AND. `q` searches product names and brands.\n\nThe **`X-Total-Count`** response header carries how many products match the filter ignoring `limit`/`offset` -- use it to render \"24 of 1,240\" and to know when to stop paging. The body stays a bare array.",
        "operationId": "list_products_products_get",
        "parameters": [
          {
            "name": "essential",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only curated Home Essentials.",
              "title": "Essential"
            },
            "description": "Only curated Home Essentials."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical category id.",
              "examples": [
                "dairy-eggs-fridge"
              ],
              "title": "Category"
            },
            "description": "Canonical category id."
          },
          {
            "name": "subcategory",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical subcategory id.",
              "examples": [
                "milk"
              ],
              "title": "Subcategory"
            },
            "description": "Canonical subcategory id."
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Canonical tag id.",
              "examples": [
                "gluten-free"
              ],
              "title": "Tag"
            },
            "description": "Canonical tag id."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search over product names.",
              "examples": [
                "milk"
              ],
              "title": "Q"
            },
            "description": "Full-text search over product names."
          },
          {
            "name": "special",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only products on special at some retailer.",
              "title": "Special"
            },
            "description": "Only products on special at some retailer."
          },
          {
            "name": "retailer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only products this retailer stocks.",
              "examples": [
                "coles"
              ],
              "title": "Retailer"
            },
            "description": "Only products this retailer stocks."
          },
          {
            "name": "multi_retailer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only products carried by 2+ retailers (comparable).",
              "title": "Multi Retailer"
            },
            "description": "Only products carried by 2+ retailers (comparable)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  },
                  "title": "Response List Products Products Get"
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "Total matches ignoring limit/offset.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/compare": {
      "post": {
        "tags": [
          "compare"
        ],
        "summary": "Price a basket as three buying strategies",
        "description": "Send the whole basket; get three complete-basket options:\n\n- `recommended-split` \u2014 cheapest practical basket using at most two retailers (a single store if a second shop does not help).\n- `cheapest-single-store` \u2014 cheapest retailer that stocks every known item.\n- `lowest-possible-price` \u2014 cheapest offer per item, any number of retailers.\n\nA strategy that cannot fulfil the complete known basket returns `total: null` and an empty `breakdown` \u2014 never a silent partial total.\n`unknown_product_ids` lists ids that do not exist at all (usually a stale basket).\n`savings` is against the most expensive complete single-store basket.\n\nCall this again whenever the basket changes.",
        "operationId": "compare_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BasketItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id",
            "examples": [
              "full-cream-milk-2l"
            ]
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Must be > 0. Removing the last unit removes the item.",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "BasketItem"
      },
...
      "CompareItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "retailer_product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retailer Product Id"
          },
          "retailer_product_name": {
            "type": "string",
            "title": "Retailer Product Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "line_total": {
            "type": "number",
            "title": "Line Total"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "retailer_product_name",
          "quantity",
          "unit_price",
          "line_total"
        ],
        "title": "CompareItem"
      },
      "CompareOption": {
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "recommended-split",
              "cheapest-single-store",
              "lowest-possible-price"
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "Complete-basket total. null when this strategy cannot fulfil every known item."
          },
          "savings": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savings",
            "description": "baseline (most expensive complete single-store) minus total. null when total or baseline is missing."
          },
          "stores": {
            "type": "integer",
            "title": "Stores",
            "description": "Number of retailer groups in breakdown."
          },
          "recommended": {
            "type": "boolean",
            "title": "Recommended"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/StoreBreakdown"
            },
            "type": "array",
            "title": "Breakdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "stores",
          "recommended",
          "breakdown"
        ],
        "title": "CompareOption"
      },
      "CompareRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BasketItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CompareRequest"
      },
      "CompareResponse": {
        "properties": {
          "options": {
            "items": {
              "$ref": "#/components/schemas/CompareOption"
            },
            "type": "array",
            "title": "Options"
          },
          "unknown_product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Product Ids",
            "description": "Ids in the request that do not exist at all.",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "options"
        ],
        "title": "CompareResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Product": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Canonical id. The basket stores this.",
            "examples": [
              "full-cream-milk-2l"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Full Cream Milk"
            ]
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand",
            "examples": [
              "Bega"
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "examples": [
              "dairy-eggs-fridge"
            ]
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory",
            "description": "null when no reliable canonical mapping exists yet.",
            "examples": [
              "milk"
            ]
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": [],
            "examples": [
              [
                "gluten-free"
              ]
            ]
          },
          "size_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Value",
            "description": "Normalised: grams, millilitres, packs or each.",
            "examples": [
              2000.0
            ]
          },
          "size_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Unit",
            "examples": [
              "ml"
            ]
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_essential": {
            "type": "boolean",
            "title": "Is Essential",
            "default": false
          },
          "min_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price",
            "description": "Cheapest current price across all retailers.",
            "examples": [
              3.1
            ]
          },
          "cheapest_retailer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cheapest Retailer",
            "description": "Which retailer sells it at `min_price`.",
            "examples": [
              "coles"
            ]
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price",
            "description": "Price per `unit_measure` at the cheapest retailer. Use with `unit_measure` to render \"$1.55 / 100g\" so a 500 g tub and a 1 kg tub compare honestly.",
            "examples": [
              1.55
            ]
          },
          "unit_measure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Measure",
            "examples": [
              "100g"
            ]
          },
          "was_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Was Price",
            "description": "Pre-special price at the cheapest retailer. null when not on special.",
            "examples": [
              4.2
            ]
          },
          "has_special": {
            "type": "boolean",
            "title": "Has Special",
            "description": "On special at at least one retailer.",
            "default": false
          },
          "retailer_count": {
            "type": "integer",
            "title": "Retailer Count",
            "description": "How many retailers stock it.",
            "default": 0,
            "examples": [
              3
            ]
          },
          "rating_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Avg",
            "description": "Retailer star rating. Sparse today -- the Woolworths detail crawl has not been run yet. Render only if present.",
            "examples": [
              4.7
            ]
          },
          "rating_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Count",
            "examples": [
              13
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Product",
        "description": "A canonical grocery -- what the user wants to buy, not one retailer's listing."
      },
...
      "StoreBreakdown": {
        "properties": {
          "retailer": {
            "type": "string",
            "enum": [
              "coles",
              "woolworths",
              "aldi",
              "harrisfarm"
            ],
            "title": "Retailer"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompareItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "retailer",
          "subtotal",
          "items"
        ],
        "title": "StoreBreakdown"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}